pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-aidefense-middleware"
version = "0.1.0"
description = "Cisco AI Defense middleware for LangChain agents"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "cisco-aidefense-sdk>=2.0.0,<3.0.0",
    "langchain>=1.2.0,<2.0.0",
    "requests>=2.32.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0,<9.0.0",
    "pytest-asyncio>=1.2.0,<2.0.0",
]
examples = [
    "langchain-openai>=1.1.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
